Removed materials.asciidoc
Manually keeping track of all the materials is too high overhead and
also not high value. Remove it and note that materials are only tracked
within each workshop write-up. TODO: May be useful to add a section later
though on big ticket items and an item inventory useful to have on hand
at all times.